feat: Improve type casting in FIR and CZT implementations for consistency

+++ src/spectral/czt.c
@@ -92,7 +92,7 @@ vv_dsp_status vv_dsp_czt_exec_cpx(
     // exponent e = +0.5 * n^2
     vv_dsp_real e = (vv_dsp_real)0.5 * (vv_dsp_real)((double)n*(double)n);
     vv_dsp_real ang = e * argW;
-    vv_dsp_real mag = (vv_dsp_real)VV_DSP_POW((double)magW, (double)e);
+    vv_dsp_real mag = (vv_dsp_real)VV_DSP_POW((double)magW, e);
     vv_dsp_cpx W_n2 = cpx_from(mag * (vv_dsp_real)VV_DSP_COS(ang), mag * (vv_dsp_real)VV_DSP_SIN(ang));
         if (n==0){ A_pow = cpx_from(1,0); A_inv_pow = cpx_from(1,0); }
         else {
@@ -106,7 +106,7 @@ vv_dsp_status vv_dsp_czt_exec_cpx(
     for (size_t k=0;k<M;++k){
         vv_dsp_real e = (vv_dsp_real)0.5 * (vv_dsp_real)((double)k*(double)k);
         vv_dsp_real ang = -e * argW;
-        vv_dsp_real mag = (vv_dsp_real)VV_DSP_POW((double)magW, (double)(-e));
+        vv_dsp_real mag = (vv_dsp_real)VV_DSP_POW((double)magW, -e);
         h[k] = cpx_from(mag * (vv_dsp_real)VV_DSP_COS(ang), mag * (vv_dsp_real)VV_DSP_SIN(ang));
     }
 
@@ -132,7 +132,7 @@ vv_dsp_status vv_dsp_czt_exec_cpx(
         vv_dsp_real dm = (vv_dsp_real)m;
         vv_dsp_real e = (vv_dsp_real)0.5 * dm * dm;
         vv_dsp_real ang = -e * argW;
-        vv_dsp_real mag = (vv_dsp_real)VV_DSP_POW((double)magW, (double)(-e));
+        vv_dsp_real mag = (vv_dsp_real)VV_DSP_POW((double)magW, -e);
         b[i] = cpx_from(mag * (vv_dsp_real)VV_DSP_COS(ang), mag * (vv_dsp_real)VV_DSP_SIN(ang));
     }
 
@@ -159,7 +159,7 @@ vv_dsp_status vv_dsp_czt_exec_cpx(
         // Final multiply by W^{+k^2/2}
         vv_dsp_real e = (vv_dsp_real)0.5 * (vv_dsp_real)(k*(double)k);
         vv_dsp_real ang = e * argW;
-        vv_dsp_real mag = (vv_dsp_real)VV_DSP_POW((double)magW, (double)(e));
+        vv_dsp_real mag = (vv_dsp_real)VV_DSP_POW((double)magW, e);
         vv_dsp_cpx wk = cpx_from(mag * VV_DSP_COS(ang), mag * VV_DSP_SIN(ang));
         vv_dsp_cpx yk = cpx_mul(a[idx], wk);
         X[k] = yk;
